Creating new common components for qtable and summary view
Created a common table component that can be reused by any table. It supports selection of an entire row and removes the select box in the quasar table.

Created a new BoatnetSummary.vue component which I'm planning to use for trips, hauls, catch, and specimen summary page (the page the user lands on when navigating to that category). Removed boatnetTrips and boatnetHauls and moved all the logic into Trips and Hauls.

Let me know what you think of this, can always revert back to the original changes, just think it makes it easier when we start building out the other screens

Also fixed up departure and return port drop down selectors.