fixing issue app config load issue when logging in for the first time (#901)
Currently all the app config is stored in pouch and we must be logged in to query the data. This was causing an issue on login page where we had no way to access config. It was also creating an issue when the app was loading and sometimes the app config would be initialized to null

To fix login issue, created a separate login config file that I'll read from instead of reading from pouch.

To fix loading issue. Initialized all appConfig values to empty objects and arrays. Also added code in App.vue to wait for the lookup docs have loaded then query pouch for appConfig info. Also modified default.vue to continue to show progress bar as we are loading appConfig info.